Type fix

+++ src/ffi.rs
@@ -16,8 +16,8 @@ extern "C" {
                      insn: &mut *const Insn)
                      -> libc::size_t;
     pub fn cs_free(insn: *const Insn, count: libc::size_t);
-    pub fn cs_reg_name(handle: csh, reg_id: libc::size_t) -> *const libc::c_char;
-    pub fn cs_insn_name(handle: csh, insn_id: libc::size_t) -> *const libc::c_char;
+    pub fn cs_reg_name(handle: csh, reg_id: libc::c_uint) -> *const libc::c_char;
+    pub fn cs_insn_name(handle: csh, insn_id: libc::c_uint) -> *const libc::c_char;
     pub fn cs_errno(handle: csh) -> CsErr;
     pub fn cs_strerror(err: CsErr) -> *const libc::c_char;
 }
